Fix double render
